Use clang as default compiler when building from Xcode.
I'm getting a lot of strange invalid access exceptions when compiling the unit tests with both gcc46 and gcc47 via macports on the development machine. Switching to clang seems to fix the problem, (probably because mp compiles Boost.Test with clang to begin with) so this seems like a good fix for now. I have seen no sign of these errors on testing machines, and unit tests do not run on either staging or deployment so this shouldn't be a problem elsewhere.